Removed UIPreference

+++ RelationshipCoach/Model/CoachingTips/CoachingTipsForFemale.swift
@@ -9,7 +9,7 @@ import Foundation
 import SwiftUI
 
 extension StoryType {
-    func coachingTipsForFemale(size: Int) -> [AttributedString] {
+    func coachingTipsForFemale(size: CGFloat) -> [AttributedString] {
         switch self {
         case .hisFault:
             return hisFaultCoachingTips(size: size)
@@ -28,33 +28,33 @@ extension StoryType {
         }
     }
     
-    func hisFaultCoachingTips(size: Int) -> [AttributedString] {
+    func hisFaultCoachingTips(size: CGFloat) -> [AttributedString] {
         var tip1: AttributedString {
             var finalString = AttributedString("")
             
             var text1 = AttributedString("Omit the words ")
             text1.foregroundColor = Color.textColor
-            text1.font = .systemFont(ofSize: CGFloat(size))
+            text1.font = .systemFont(ofSize: size)
             finalString += text1
             
             var text2 = AttributedString("You, Because, Hurt")
             text2.foregroundColor = Gender.female.color
-            text2.font = .systemFont(ofSize: CGFloat(size), weight: .bold)
+            text2.font = .systemFont(ofSize: size, weight: .bold)
             finalString += text2
             
             var text3 = AttributedString(" and ")
             text3.foregroundColor = Color.textColor
-            text3.font = .systemFont(ofSize: CGFloat(size))
+            text3.font = .systemFont(ofSize: size)
             finalString += text3
             
             var text4 = AttributedString("Disappointed")
             text4.foregroundColor = Gender.female.color
-            text4.font = .systemFont(ofSize: CGFloat(size), weight: .bold)
+            text4.font = .systemFont(ofSize: size, weight: .bold)
             finalString += text4
             
             var text5 = AttributedString(" when you're expressing your feelings about a situation, so he doesn't feel blamed or shamed.")
             text5.foregroundColor = Color.textColor
-            text5.font = .systemFont(ofSize: CGFloat(size))
+            text5.font = .systemFont(ofSize: size)
             finalString += text5
             
             return finalString
@@ -69,12 +69,12 @@ extension StoryType {
             
             var text1 = AttributedString("To do this, figure out the result from the situation; ex: If he ignored you, figure out the result of being ignored. \"I felt alone when that happened\" ")
             text1.foregroundColor = Color.textColor
-            text1.font = .systemFont(ofSize: CGFloat(size))
+            text1.font = .systemFont(ofSize: size)
             finalString += text1
             
             var text2 = AttributedString("(instead of \"you ignored me\")")
             text2.foregroundColor = Gender.female.color
-            text2.font = .systemFont(ofSize: CGFloat(size), weight: .bold)
+            text2.font = .systemFont(ofSize: size, weight: .bold)
             finalString += text2
             return finalString
         }
@@ -106,7 +106,7 @@ extension StoryType {
         return [tip1, tip2, tip3, tip4, tip5, tip6, tip7, tip8, tip9]
     }
     
-    func myFaultCoachingTipsForFemale(size: Int) -> [AttributedString] {
+    func myFaultCoachingTipsForFemale(size: CGFloat) -> [AttributedString] {
         var tip1: AttributedString {
             "Notice that she never blamed, shamed, felt entitled or was defensive when she apologized."
         }
@@ -125,7 +125,7 @@ extension StoryType {
         return [tip1, tip2, tip3, tip4]
     }
     
-    func toxicChatCoachingTipsForFemale(size: Int) -> [AttributedString] {
+    func toxicChatCoachingTipsForFemale(size: CGFloat) -> [AttributedString] {
         var tip1: AttributedString {
             "The goal is to only focus on your feelings (self-centered in a good way). You never told him how he should or should not feel, only how YOU felt."
         }
@@ -166,7 +166,7 @@ extension StoryType {
         return [tip1, tip2, tip3, tip4, tip5, tip6, tip7, tip8, tip9, tip10]
     }
     
-    func receivingTalkCoachingTips(size: Int) -> [AttributedString] {
+    func receivingTalkCoachingTips(size: CGFloat) -> [AttributedString] {
         var tip1: AttributedString {
             "Men need to give and have their gift received. They also need appreciation, admiration, respect and love, but most of all, respect."
         }
@@ -204,7 +204,7 @@ extension StoryType {
         return [tip1, tip2, tip3, tip4, tip5, tip6, tip7, tip8, tip9]
     }
     
-    func sweetTalkCoachingTipsForFemale(size: Int) -> [AttributedString] {
+    func sweetTalkCoachingTipsForFemale(size: CGFloat) -> [AttributedString] {
         
         
         var tip1: AttributedString {
